refactor_message_adapter Improve the helpers
